sample_attack/requirements.txt
PyYAML==5.4.1
easydict==1.9

scipy>=0.14.0

# For robustbench.
tqdm>=4.56.1 
requests>=2.25.1
pandas>=1.2.4
numpy>=1.19.4